index.html
Iam not sure I understand the review changes page. It says I have a body tag in 2 places there is no body tag both in lines they should not be in. I believe the title and head are children of the html tag and siblings of the body tag  and I believe it is correct as I typed it.